+++ src/onegov.core/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
+- Adds the ability to run transactions in read only mode.
+  [href]
+
 - Switch postgres isolation level to SERIALIZABLE for all sessions.
   [href]